Feature TypeFormWidget: Embedding typeform inside of my react application, currently not working but shows on page and isn't directing to the actual typeform.
Figured this would be an easy addition and quick form for sending emails to me on my portfolio until I build the backend up.
I removed the standard HTML form component, and decided on Typeform as it's easily configurable and is easy to fit with the style of this site whilst allowing for integration with my email.